+++ anxcloud/schema_virtual_server.go
@@ -60,6 +60,7 @@ func schemaVirtualServer() map[string]*schema.Schema {
 		"sockets": {
 			Type:     schema.TypeInt,
 			Optional: true,
+			Computed: true,
 			Description: "Amount of CPU sockets Number of cores have to be a multiple of sockets, as they will be spread evenly across all sockets. " +
 				"Defaults to number of cores, i.e. one socket per CPU core.",
 		},